+++ crates/module-system/sov-modules-macros/tests/integration/trybuild/rpc/expose_rpc_associated_type_not_static.rs
@@ -25,6 +25,7 @@ pub trait Data:
     + serde::de::DeserializeOwned
     + borsh::BorshSerialize
     + borsh::BorshDeserialize
+    + schemars::JsonSchema
     + Default
 {
 }